refactor: 추구미 유형 선택 컴포넌트명 수정 및 파일 구조 변경
