+++ src/router/index.js
@@ -4,6 +4,7 @@ import { useAuthStore } from '@/stores/authStore';
 import AssetConnectView from '@/views/asset/AssetConnectView.vue';
 import AssetSelectView from '@/views/asset/AssetSelectView.vue';
 import AssetReportView from '@/views/asset/report/AssetReportView.vue';
+import ChoogooMiSelectView from '@/views/choogoomi/ChoogooMiSelectView.vue';
 import HomeView from '@/views/home/HomeView.vue';
 import TransactionView from '@/views/home/TransactionView.vue';
 import LoginView from '@/views/login/LoginView.vue';
@@ -14,7 +15,6 @@ import MyPageEditInfoView from '@/views/mypage/MyPageEditInfoView.vue';
 import MyPageRecordView from '@/views/mypage/MyPageRecordView.vue';
 import MyPageView from '@/views/mypage/MyPageView.vue';
 import RankingView from '@/views/ranking/RankingView.vue';
-import CharacterSelectComponent from '@/views/signup/components/character/CharacterSelectComponent.vue';
 import SurveyView from '@/views/signup/components/survey/SurveyView.vue';
 import SignupView from '@/views/signup/SignupView.vue';
 
@@ -62,9 +62,9 @@ const router = createRouter({
       component: AssetReportView,
     },
     {
-      path: '/character',
-      name: 'character',
-      component: CharacterSelectComponent,
+      path: '/choogoomi',
+      name: 'choogoomi',
+      component: ChoogooMiSelectView,
     },
     {
       path: '/ranking',
