+++ src/views/choogoomi/ChoogooMiCard.vue
