+++ src/views/choogoomi/ChoogooMiDetailModal.vue
